+++ components/qr-code/demo/errorlevel.tsx
@@ -6,10 +6,10 @@
 import React, { useState } from 'react';
 import { Radio, QrCode } from 'antd';
 
-export default () => {
+const App: React.FC = () => {
   const [level, setLevel] = useState<string>('L');
   return (
-    <div>
+    <>
       <Radio.Group value={level} onChange={(e) => setLevel(e.target.value)}>
         <Radio.Button value="L">L</Radio.Button>
         <Radio.Button value="M">M</Radio.Button>
@@ -17,10 +17,13 @@ export default () => {
         <Radio.Button value="H">H</Radio.Button>
       </Radio.Group>
       <QrCode
+        style={{ marginTop: 16 }}
         errorLevel={level}
-        value="https://hitu.alipay.com/packages/@alipay/tech-ui/home"
-        logo="https://gw-office.alipayobjects.com/basement_prod/c83c53ab-515e-43e2-85d0-4d0da16f11ef.svg"
+        value="https://ant.design/"
+        logo="https://gw.alipayobjects.com/zos/antfincdn/%24C9tmj978R/Carousel.svg"
       />
-    </div>
+    </>
   );
 };
+
+export default App;
