+++ components/qr-code/demo/popover.tsx
@@ -1,16 +1,13 @@
-/**
- * uuid: b83bf277
- * title: popover 带 logo 例子
- */
-
 import React from 'react';
 import { QrCode } from 'antd';
 
-export default () => (
+const App: React.FC = () => (
   <QrCode
-    value="http://www.tmall.com"
-    logo="https://gw-office.alipayobjects.com/basement_prod/c83c53ab-515e-43e2-85d0-4d0da16f11ef.svg"
+    value="https://ant.design/"
+    logo="https://gw.alipayobjects.com/zos/antfincdn/%24C9tmj978R/Carousel.svg"
     logoSize={64}
     popover
   />
 );
+
+export default App;
