+++ components/qr-code/demo/rendermode.tsx
@@ -1,15 +1,12 @@
-/**
- * uuid: f54dc6af
- * title: 渲染为canvas
- */
-
 import React from 'react';
 import { QrCode } from 'antd';
 
-export default () => (
+const App: React.FC = () => (
   <QrCode
     mode="svg"
-    value="http://www.baidu.com"
-    logo="https://gw-office.alipayobjects.com/basement_prod/c83c53ab-515e-43e2-85d0-4d0da16f11ef.svg"
+    value="https://ant.design/"
+    logo="https://gw.alipayobjects.com/zos/antfincdn/%24C9tmj978R/Carousel.svg"
   />
 );
+
+export default App;
