--- components/qr-code/demo/service.md
@@ -1,7 +1,0 @@
-## zh-CN
-
-`format` 属性指定格式。
-
-## en-US
-
-You can set a custom text by setting the `format` prop.
