+++ components/qr-code/demo/withoutlogo.tsx
@@ -1,9 +1,6 @@
-/**
- * uuid: 0a9254ef
- * title: 不带 logo 例子
- */
-
 import React from 'react';
 import { QrCode } from 'antd';
 
-export default () => <QrCode value="http://www.tmall.com" />;
+const App: React.FC = () => <QrCode value="https://ant.design/" />;
+
+export default App;
